rewrite to uses task loops

+++ temp-chat-beta.py
@@ -4,6 +4,8 @@ from datetime import datetime, timedelta
 import os
 import logging
 
+from discord.ext import tasks, commands
+
 intents = discord.Intents.default()
 intents.message_content = True
 client = discord.Client(intents=intents)
@@ -12,28 +14,32 @@ client = discord.Client(intents=intents)
 token = os.getenv("BOT_TOKEN")
 temp_chat_beta = int(os.getenv("CHANNEL_ID"))
 delay = int(os.getenv("BEFORE_MINUTES", default="60"))
-sleep_time = int(os.getenv("SLEEP_TIME", default="30"))
+loop_time = int(os.getenv("SLEEP_TIME", default="30"))
+sleep_time = float(os.getenv("DELETE_SLEEP", default="0.25"))
 
-async def retrieve_messages():
+@tasks.loop(seconds=loop_time)
+async def delete_old_messages():
+    try:
+        channel = client.get_channel(temp_chat_beta)
+        before_time = datetime.now() - timedelta(minutes=delay)
+        messages = channel.history(before=before_time)
 
-    channel = client.get_channel(temp_chat_beta)
-    before_time = datetime.now() - timedelta(minutes=delay)
-    messages = channel.history(before=before_time)
-    messages_len = 0
-    async for message in messages:
-        await message.delete()
-        messages_len += 1
-    if messages_len > 0:
-        logging.info(f"deleted {messages_len} messages")
+        if messages:
+            async for message in messages:
+                await message.delete()
+            await asyncio.sleep(sleep_time)
 
+    except Exception as e:
+        print(f"An error occurred {e}")
 
-async def main():
-    while True:
-        await retrieve_messages()
-        await asyncio.sleep(sleep_time)
+@delete_old_messages.before_loop
+async def startup():
+    print("Waiting until the client is ready")
+    await client.wait_until_ready()
 
 @client.event
 async def on_ready():
-    await main()
+    print(f"Logged in as {client.user}")
+    delete_old_messages.start()
 
 client.run(token=token)
